requirements.txt
Django==1.3.1
feedparser==5.1
gunicorn==0.12.2
python-dateutil==1.5
python-sunlightapi==1.1.0
django-social-auth==0.7.25
Pillow
